rename pharos directory to pharos-api-server
